Movement, movement!

+++ src/org/kruithne/gamething/rendering/RenderPositionalObject.java
@@ -2,34 +2,34 @@ package org.kruithne.gamething.rendering;
 
 public abstract class RenderPositionalObject implements IRenderable
 {
-	public RenderPositionalObject(int x, int y)
+	public RenderPositionalObject(float x, float y)
 	{
 		drawX = x;
 		drawY = y;
 	}
 
-	public RenderPositionalObject setDrawX(int x)
+	public RenderPositionalObject setDrawX(float x)
 	{
 		drawX = x;
 		return this;
 	}
 
-	public RenderPositionalObject setDrawY(int y)
+	public RenderPositionalObject setDrawY(float y)
 	{
 		drawY = y;
 		return this;
 	}
 
-	public int getDrawX()
+	public float getDrawX()
 	{
 		return drawX;
 	}
 
-	public int getDrawY()
+	public float getDrawY()
 	{
 		return drawY;
 	}
 
-	protected int drawX;
-	protected int drawY;
+	protected float drawX;
+	protected float drawY;
 }
